refactor(pd): split worker into decode/prefill handlers and slim trace
Split worker.py (1000+ lines) into:
- decode_worker.py: D-side RDMA receive + prefill dispatch
- prefill_worker.py: P-side KV push + async RDMA send/finalize
- worker.py: thin facade composing both handlers

Shrink _PushTrace (12 → 5 fields), _PushFinalizeTask (12 → 8 fields),
_LayerPushTask (7 → 4 fields). Remove per-chunk duplicate metrics and
per-layer INFO log spam. Move block range coalescing to layout.py.
Replace OrderedDict LRU slot_mapping cache with single-value cache.
